Move shared components out of the Embed namespace
We don't really need an Embed namespace because this is an application and not a library.  The extra namespace is just clutter
