+++ app/components/content_not_available_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class ContentNotAvailableComponent < ViewComponent::Base
+end
