--- app/components/embed/content_not_available_component.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Embed
-  class ContentNotAvailableComponent < ViewComponent::Base
-  end
-end
