+++ app/components/embed/media_component.html.erb
@@ -1,4 +1,4 @@
-<%= render Embed::CompanionWindowsComponent.new(viewer:, stimulus_controller: 'media-tag transcript') do |component| %>
+<%= render ::CompanionWindowsComponent.new(viewer:, stimulus_controller: 'media-tag transcript') do |component| %>
   <% component.with_header_button do %>
     <button data-action="click->companion-window#displayAccessibility" aria-label="Accessibility" aria-controls="accessibility">
       <span class="button-label"><svg class="MuiSvgIcon-root" focusable="false" aria-hidden="true" viewBox="0 0 24 24"><path d="M12 2c1.1 0 2 .9 2 2s-.9 2-2 2-2-.9-2-2 .9-2 2-2zm9 7h-6v13h-2v-6h-2v6H9V9H3V7h18v2z"></path></svg></span>
@@ -71,7 +71,7 @@
         </div>
       </div>
       <div data-auth-restriction-target="noAccess" hidden>
-         <%= render Embed::ContentNotAvailableComponent.new %>
+         <%= render ::ContentNotAvailableComponent.new %>
       </div>
     </div>
   <% end %>
