+++ app/components/embed/pdf_component.html.erb
@@ -1,4 +1,4 @@
-<%= render Embed::CompanionWindowsComponent.new(viewer:, stimulus_controller: 'file-auth fullscreen') do |component| %>
+<%= render ::CompanionWindowsComponent.new(viewer:, stimulus_controller: 'file-auth fullscreen') do |component| %>
   <% component.with_header_button do %>
     <%# chromium has a bug preventing the esc key from exiting full screen in some situations, see
         https://issues.chromium.org/issues/40131988 and https://github.com/sul-dlss/sul-embed/issues/2162 %>
@@ -21,7 +21,7 @@
             data-action="iiif-manifest-received@window->file-auth#parseFiles auth-success@window->pdf#show">
           </div>
     <% else %>
-       <%= render Embed::ContentNotAvailableComponent.new %>
+       <%= render ::ContentNotAvailableComponent.new %>
     <% end %>
     </div>
   <% end %>
