+++ app/components/legacy/download/geo_component.html.erb
@@ -2,7 +2,7 @@
   <%= render Legacy::DownloadPanelComponent.new do %>
     <% viewer.purl_object.contents.each do |resource| %>
       <ul class='sul-embed-download-list'>
-        <%= render Embed::Download::FileListItemComponent.with_collection(resource.files, prefer_filename: true) %>
+        <%= render Download::FileListItemComponent.with_collection(resource.files, prefer_filename: true) %>
       </ul>
     <% end %>
   <% end %>
